Use singular form for files and names
