+++ src/httpProblemDetailDefinition.ts
@@ -1,6 +1,6 @@
 import { URI, HttpStatusCode } from './types';
 
-export interface HttpProblemDetailsDefinition {
+export interface HttpProblemDetailDefinition {
   type: URI;
   title: string;
   status: HttpStatusCode;
